Add a Hypothesis to the Project

Adding an "arm" to the project reflects a series of hypothesis generated around the core data that was set in the preceding function. The hypothesis formula allows for modifiers to both the formula itself, as well as how it should be applied to the data set.

add_hypothesis(
  project,
  name,
  formula,
  combination = "direct",
  test,
  which_data = NULL,
  .strata = NA,
  .stage = "hypothesis",
  ...
)

Arguments

project
Object of class project

name
Character string to identify this arm. Must be unique and not previously specified.

formula
Formula showing relationship of outcomes and predictors, and is essentially the hypothesis. It allows for complex formulas e.g. multiple predictors, multiple exposures, and multiple covariates, that can be organized into individual formulas based on additional modifiers. Each RHS term of the formula is considered an outcome variable, and is analyzed as a single outcome. Each LHS term is consider a predictor, and can be modified as below:

exp() is placed around a term to define as an independent exposure, which will be placed in separate formulas from any other term marked as an exposure

cov() is placed around a term for any predictors that should be maintained/fixed in all models, which can include complex terms, such as mixed effects

Of note, these formula modifiers do not apply to htest objects yet.

For example, the equation below describes two independent exposures "x1" and "x2" that should be conditionally regressed for every level of "z"

$$y ~ exp(x1) + exp(x2) + x3 + x4 + cov((1 | z))$$

combination
The building pattern for how to put together the overall plan. It defines variable relationships that will be used. The options for the combination currently include:

direct will define the relationship as y ~ x

sequential will define the relationship as y ~ x1, y ~ x1 + x2

parallel will define the relationship as y ~ x1, y ~ x2

test
The testing that will be used, defined as how the analysis or hypothesis testing should be performed. There are several options on which specification to use, anywhere from regression modeling to inferential statistics. The test applies to the collection of hypotheses.

A model specified by the parsnip package, which includes the mode and #' computational engine

A statistical test, such as a t.test, which may require additional parameters. These can be given as additional, unmatched arguments. This option currently supports only hypothesis tests, of class htest.

which_data
Name of "row" or dataset that should be analyzed. Defaults to the most recent dataset added.

.strata
How the data should be split or stratified. References the name of the data given in set_data() that the models will be with fit against, splitting the data into subsets. This helps to perform hypothesis testing on subsets or strata of the data. It defaults to NULL (which means the full data will be used) experimental

.stage
Internal marker of project status.

...
This should reflect the additional parameters that may need to be given to the test argument, such as paired = TRUE for t.test(). The additional parameters must be named to allow them to be passed successfully.

Value

project object