Create a Hypothesis

hypothesize() creates a hypothesis object that can be placed into a modeling framework. This framework allows for multiple hypotheses to be modeled, and allows for assessment of confounding variables.

hypothesize(h, ...)

# S3 method for formula
hypothesize(h, combination, test, test_opts = NULL, data, strata = NULL, ...)

# S3 method for default
hypothesize(h, ...)

Arguments

h
General variable that can be cast into a hypothesis object. Currently this accepts the following object types:

formula objects, which be given additional modifiers as described. Please see expand_formula() or the formula section below for details.

...
For additional variables based on the generic method invoked.

combination
The building pattern for how to put together the overall plan. It defines variable relationships that will be used. The options for the combination currently include:

direct will define the relationship as y ~ x

sequential will define the relationship as y ~ x1, y ~ x1 + x2

parallel will define the relationship as y ~ x1, y ~ x2

test
The testing that will be used, defined as how the analysis or hypothesis testing should be performed. There are several options on which specification to use, anywhere from regression modeling to inferential statistics. The test applies to the collection of hypotheses.

A model specified by the parsnip package, which includes the mode and computational engine

A statistical test, such as a t.test, which may require additional parameters. These can be given as additional, unmatched arguments. This option currently supports only hypothesis tests, of class htest.

test_opts
For optional or additional parameters to be given to the test argument. Should be in format of a list with name-pairs, such as list(name = argument) pairs.

data
Data set to be linked to this hypothesis. It assumes data is one of type c("tbl_df", "tbl", "data.frame"). The name of the data set is taken from the object itself.

strata
How the data should be split or stratified. References the name of the data given in that the models will be with fit against, splitting the data into subsets. This helps to perform hypothesis testing on subsets or strata of the data. It defaults to NULL (which means the full data will be used). This argument will only be incorporated if data is present.

Value

An object of class hypothesis

Warning

The formula object does not necessarily work for htest objects as it is still experimental.

Formulae

When supplying a formula object, each RHS term of the formula is considered an outcome variable, and is analyzed as a single outcome. Each LHS term is consider a predictor, and can be modified as below:

X() is placed around a term to define as an independent exposure, which will be placed in separate formulae from any other term marked as an exposure

F() is placed around a term for any predictors that should be maintained/fixed in all models, which can include complex terms, such as mixed effects

For example, the equation below describes two independent exposures "x1" and "x2" that should be conditionally tested for every level of "z"

$$y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))$$