Retrieve, Modify, and Destroy Components From framework Object

These functions are essentially helper functions to simplify the manipulation of framework objects. They may retrieve or modify or destroy components of the framework object as described.

Retrieval functions:

get_data() retrieves the data set associated with a specified hypothesis and data name

get_test() retrieves the test associated with a hypothesis

get_formulae() retrieves the list of formulaes associated with a hypothesis and data set.

Reporting functions

report_stage() creates a message to identify at what stage a hypothesis has been developed to.

Update functions:

update_status() modifies the components of a framework, usually used internally to help update the stage/status of the specified object

get_data(x, name)

get_test(x, name)

get_combination(x, name)

get_formulae(x, name)

get_parameters(x, name)

update_status(x, name, stage = NA, run = FALSE)

Arguments

x
framework object

name
Name of hypothesis object to pull components from

...
For optional parameters

Value

The get_*() functions return the named object from the framework. The update_*() functions return the entire framework after modification.